Move error generator to utilities
