+++ tracker/core/developer-tools/package.json
@@ -42,7 +42,6 @@
     }
   },
   "scripts": {
-    "generate": "ts-node src/scripts/generateErrorMessages.ts",
     "build": "tsup src/index.ts --format cjs,esm --legacy-output --minify --sourcemap --clean",
     "prettify": "prettier --write .",
     "tsc": "tsc --noEmit",
@@ -58,7 +57,6 @@
     "jest-standard-reporter": "^2.0.0",
     "prettier": "^2.5.1",
     "ts-jest": "^27.1.3",
-    "ts-node": "^10.7.0",
     "tsup": "^5.12.0",
     "typescript": "^4.6.2"
   },
