+++ tracker/core/tracker/src/index.ts
@@ -16,6 +16,7 @@ export * from './generated/EventNames';
 export * from './cleanObjectFromInternalProperties';
 export * from './Context';
 export * from './ContextValidationRules';
+export * from './ErrorMessages';
 export * from './EventRecorderInterface';
 export * from './helpers';
 export * from './LocationTreeInterface';
@@ -24,6 +25,7 @@ export * from './Tracker';
 export * from './TrackerConsoleInterface';
 export * from './TrackerDeveloperToolsInterface';
 export * from './TrackerEvent';
+export * from './TrackerPlatform';
 export * from './TrackerPluginInterface';
 export * from './TrackerPluginLifecycleInterface';
 export * from './TrackerPlugins';
