+++ tracker/core/utilities/package.json
@@ -5,15 +5,21 @@
   "license": "Apache-2.0",
   "description": "Objectiv Core Utilities",
   "files": [
-    "src/generator.js"
+    "src/generateSchema.js",
+    "src/generateErrorMessages.ts"
   ],
   "scripts": {
     "prettify": "prettier --write .",
-    "generate": "(cd src && node generator.js)",
-    "depcheck": "npx depcheck"
+    "generate:schema": "(cd src && node generateSchema.js)",
+    "generate:errors": "tsup src/generateErrorMessages.ts --onSuccess 'node dist/generateErrorMessages.js'",
+    "generate": "yarn generate:schema && yarn generate:errors",
+    "depcheck": "npx depcheck --ignores uuid"
   },
   "devDependencies": {
+    "@objectiv/tracker-core": "^0.0.22",
     "json5": "^2.2.0",
-    "prettier": "^2.5.1"
+    "prettier": "^2.5.1",
+    "tsup": "^5.12.0",
+    "typescript": "^4.6.2"
   }
 }
